feat(router): moved the router into its own separate folder
