+++ backend/index.js
@@ -2,7 +2,7 @@ require('dotenv').config();
 const express = require('express');
 const mongoose = require('mongoose');
 const cors = require('cors');
-const router = require('./router');
+const router = require('./routes/router');
 
 const app = express();
 const PORT = process.env.SERVER_PORT || 8080;
