+++ backend/routes/router.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const controller = require('./controller');
+const controller = require('../controller');
 
 router.get('/main-canvas', controller.getMainCanvas);
 router.get('/canvas/:_id', controller.getCanvas);
